# docs/swagger.yaml
definitions:
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  model.Comment:
    properties:
      comment_message:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo:
        $ref: '#/definitions/model.Photo'
      photo_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Photo:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_caption:
        example: photo caption
        type: string
      photo_title:
        example: photo title
        type: string
      photo_url:
        example: photo url
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.SwagComment:
    properties:
      comment_message:
        example: Insert you message here
        type: string
    type: object
...
  model.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photos:
        items:
          $ref: '#/definitions/model.Photo'
        type: array
      updated_at:
        type: string
      user_age:
        type: integer
      user_email:
        type: string
      user_password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /comments/{id}:
    delete:
      description: |-
        Return forbidden if the user is not the owner of the comment
        Validates: (Message)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Delete certain comment that is specified by the id
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Input valid comment id through parameter
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comment'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Get comment from specific comment id
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: |-
        Return forbidden if the user is not the owner of the comment
        Validates: (Message)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.SwagComment'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Update certain comment that is specified by the id
      tags:
      - Comment
  /comments/photo/{id}:
    get:
      description: Input valid photo ID in the parameter
      parameters:
      - description: Photo ID here
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comment'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Return all comments from specific photo ID
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: |-
        post a json data to comment
        Validates: (Message)
      parameters:
      - description: Photo ID here
        in: path
        name: id
        required: true
        type: integer
      - description: '-'
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.SwagComment'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Add comment to a photo
      tags:
      - Comment
...
swagger: "2.0"